requirements.txt
discord.py-self
discum
asyncio
bcrypt
requests
discord_rpc.py
ua_parser